vkVideoDecoder: Fix data conversion issue

+++ vk_video_decoder/demos/vk-video-dec/Main.cpp
@@ -125,7 +125,7 @@ int main(int argc, const char **argv)
             return EXIT_FAILURE;
         }
         assert(displayShell->PhysDeviceCanPresent(vkDevCtxt.getPhysicalDevice(),
-                                                  vkDevCtxt.GetPresentQueueFamilyIdx()));
+                                                  (uint32_t)vkDevCtxt.GetPresentQueueFamilyIdx()));
 
         vkDevCtxt.CreateVulkanDevice(numDecodeQueues,           // numDecodeQueues
                                      0,                         // num encode queues
@@ -249,7 +249,7 @@ int main(int argc, const char **argv)
         }
 
         VkSharedBaseObj<VkVideoQueue<VulkanDecodedFrame>> videoQueue(vulkanVideoProcessor);
-        DecoderFrameProcessorState frameProcessor(&vkDevCtxt, videoQueue, decoderConfig.decoderQueueSize);
+        DecoderFrameProcessorState frameProcessor(&vkDevCtxt, videoQueue, (int32_t)decoderConfig.decoderQueueSize);
 
         bool continueLoop = true;
         do {
