+++ vk_video_decoder/test/vulkan-video-simple-dec/Main.cpp
@@ -92,7 +92,7 @@ static bool GetNextFrame(VkSharedBaseObj<VulkanVideoDecoder>& vulkanVideoDecoder
     }
 
     if (gotFrame) {
-        curFrameDataQueueIndex = (curFrameDataQueueIndex + 1) % frameDataQueue.size();
+        curFrameDataQueueIndex = (curFrameDataQueueIndex + 1) % (uint32_t)frameDataQueue.size();
     }
 
     return continueLoop;
